+++ install.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-docker build --rm --no-cache -t app-boilerplate:latest -f Dockerfile .
+docker build --rm --no-cache -t coordinate-extract-app:latest -f Dockerfile .